+++ CHANGELOG.md
@@ -7,7 +7,8 @@
     - Main library now lives under `lib/` in source control
     - New examples module under `examples/` in source control
 - CLI Improvements
-    - User defined command aliases are now supported
+    - User defined command
+    -  aliases are now supported
     - CLI builder classes are now public and have their own `builder` package
     - Parser Improvements
         - All parser errors in the `parser` package have public constructors
@@ -30,6 +31,7 @@
         - Basic ANSI Colors (8 Colors, normal and bright variants)
         - 256 Colors (Basic ANSI Colors plus 3 colour palettes plus grayscale palette) - See [color chart](https://camo.githubusercontent.com/6378594a85c578517c5a4e494789bd4d66c9e46b/68747470733a2f2f7261772e6769746875622e636f6d2f666f697a652f676f2e7367722f6d61737465722f787465726d5f636f6c6f725f63686172742e706e67) for more detail
         - True Color (24 bit colour i.e. 16 million colours) - Many terminals may not support this mode
+     - Support for basic text decorated output streams and writers e.g. bold, underline
 
 ## 0.9.2